Remove EntitySchema's default definition. Use definition's merge/construct
Entity schema's used to have a default definition of MapSchema.
This made sense in most cases but it was a bit magic. It also put the
burden of knowledge on the user to know that if they wanted an Entity of
a different type (array/value) they needed to override the definition.

This way is a little more verbose, but a lot more clear.