test: split Api tests from createActionTests, test observables
Most of the Api tests were action tests. This splits them out so that
the api tests are just shape replated and the action tests describe
the dispatching of redux actions.

Also changed the action tests to focus on what goes in and what
ends up being dispatched. The previous tests were a little low level and
harder to understand.