fix: request hook breaks when using observable
createRequestAction now returns a promise to RequestHook even when observables are used,
so request hook can call .then() on the result

An alternative is to handle the promise vs observable cases differently in RequestHook,
but this gives react-enty potentially unnecessary knowledge that observables exist,
when a promise is sufficient to handle the timing of when data is received.
If this alternative is pursued, then the isObservable function would need to be shared between enty-state and react-enty.